Separate out function utils
